\name{treePlotter-package}
\alias{treePlotter}
\docType{package}
\title{
\packageTitle{treePlotter}
}
\description{
\packageDescription{treePlotter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{treePlotter}
\packageIndices{treePlotter}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{treePlotter}

Maintainer: \packageMaintainer{treePlotter}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
